refactor(cketh): put the account cbor encoding with the other cbor helpers
The account encoding is not about attestations: it is how the minter's events
carry an account, next to the id encoding. A roundtrip proptest covers what
the shape has to survive — a principal of any length up to 29 bytes and a
subaccount that is either absent or exactly 32.
