+++ rs/ethereum/cketh/minter/src/cbor/mod.rs
@@ -1,3 +1,4 @@
+pub mod account;
 pub mod id;
 
 #[cfg(test)]
